feat(web): separate `Calendar` from component, add `ScheduleItem`
