+++ apps/web/src/features/project-schedule/ui/index.ts
@@ -1,2 +1,3 @@
 export { default as EntireSchedule } from './EntireSchedule';
 export { default as EntireTodo } from './EntireTodo';
+export { default as ScheduleCalendar } from './ScheduleCalendar';
