+++ apps/web/src/widgets/project-schedule/ui/ScheduleContainer.tsx
@@ -6,7 +6,7 @@ export default function ScheduleContainer() {
   return (
     <div className="scrollbar-hide relative flex size-full flex-col overflow-scroll">
       <Breadcrumb />
-      <div className="scrollbar-hide gap-y-15 flex h-screen w-full flex-col overflow-scroll px-10 pt-20">
+      <div className="scrollbar-hide gap-y-15 pt-30 flex w-full flex-col overflow-scroll px-10">
         <EntireSchedule />
         <EntireTodo />
       </div>
